fix(naval): split fleet tile widget to satisfy panel size lint
Extract FleetExpansionTile from naval_units_panel.dart so the panel file remains within the repo-lint 700-line limit without changing behavior.

Made-with: Cursor
